requirements.txt
discord.py==2.4.0
apscheduler==3.10.4
web3==6.11.3
python-dotenv==1.0.1
